+++ e2e/runner/setup_solana.go
@@ -180,9 +180,7 @@ func (r *E2ERunner) ensureSolanaChainParams() error {
 		},
 	}
 
-	updateMsg := observertypes.NewMsgUpdateChainParams(creator, chainParams)
-
-	if _, err := r.ZetaTxServer.BroadcastTx(utils.OperationalPolicyName, updateMsg); err != nil {
+	if err := r.ZetaTxServer.UpdateChainParams(chainParams); err != nil {
 		return errors.Wrap(err, "unable to broadcast solana chain params tx")
 	}