+++ x/crosschain/types/expected_keepers.go
@@ -151,7 +151,7 @@ type FungibleKeeper interface {
 		protocolContractVersion ProtocolContractVersion,
 		isCrossChainCall bool,
 	) (*evmtypes.MsgEthereumTxResponse, bool, error)
-	ProcessV2RevertDeposit(
+	ProcessRevert(
 		ctx sdk.Context,
 		inboundSender string,
 		amount *big.Int,
@@ -160,6 +160,17 @@ type FungibleKeeper interface {
 		asset string,
 		revertAddress ethcommon.Address,
 		callOnRevert bool,
+		revertMessage []byte,
+	) (*evmtypes.MsgEthereumTxResponse, error)
+	ProcessAbort(
+		ctx sdk.Context,
+		inboundSender string,
+		amount *big.Int,
+		outgoing bool,
+		chainID int64,
+		coinType coin.CoinType,
+		asset string,
+		abortAddress ethcommon.Address,
 		revertMessage []byte,
 	) (*evmtypes.MsgEthereumTxResponse, error)
 	CallUniswapV2RouterSwapExactTokensForTokens(