+++ changelog.md
@@ -4,13 +4,17 @@
 
 ### Features
 
+* [3414](https://github.com/zeta-chain/node/pull/3414) - support advanced abort workflow (onAbort)
 * [3461](https://github.com/zeta-chain/node/pull/3461) - add new `ConfirmationParams` field to chain params to enable multiple confirmation count values, deprecating `confirmation_count`
 * [3489](https://github.com/zeta-chain/node/pull/3489) - add Sui chain info
 * [3455](https://github.com/zeta-chain/node/pull/3455) - add `track-cctx` command to zetatools
 * [3506](https://github.com/zeta-chain/node/pull/3506) - define `ConfirmationMode` enum and add it to `InboundParams`, `OutboundParams`, `MsgVoteInbound` and `MsgVoteOutbound`
 * [3469](https://github.com/zeta-chain/node/pull/3469) - add `MsgRemoveInboundTracker` to remove inbound trackers. This message can be triggered by the emergency policy.
 * [3450](https://github.com/zeta-chain/node/pull/3450) - SOL withdraw and call integration
-* [3520] (https://github.com/zeta-chain/node/pull/3520) - SPL withdraw and call integration
+* [3538](https://github.com/zeta-chain/node/pull/3538) - implement `MsgUpdateOperationalChainParams` for updating operational-related chain params with operational policy
+* [3534] (https://github.com/zeta-chain/node/pull/3534) - Add Sui deposit & depositAndCall
+* [3541](https://github.com/zeta-chain/node/pull/3541) - implement `MsgUpdateZRC20Name` to update the name or symbol of a ZRC20 token
+* [3520](https://github.com/zeta-chain/node/pull/3520) - SPL withdraw and call integration
 * [3527](https://github.com/zeta-chain/node/pull/3527) - SOL/SPL withdraw and call revert integration
 
 ### Refactor